// jm-next-app/src/types/JobPostType.ts
export type JobPost = {
    currency: string | null;
    company_logo: string | null;
    company_addresses: string | null;
    min_amount: number | null;
    location: string | null;
    job_url_direct: string | null;
    description: string | null;
    id: string | null;
    listing_type: string;
    _id: string;
    date_posted: string;
    emails: string;
    company_url_direct: string;
    extracted_keywords: string[];
    max_amount: number | null;
    company_url: string | null;
    company_industry: string | null;
    company_description: string | null;
    interval: string | null;
    is_remote: boolean | null;
    title: string;
    company_num_employees: string | null;
    salary_source: string | null;
    job_url: string;
    site: string;
    job_function: string | null;
    company: string | null;
    job_level: string | null;
    job_type: string | null;
    company_revenue: string | null;
}